:bug: Break up ProfileSheetHeaderView
Should address the "the compiler is unable to type-check this expression
in reasonable time" issue.
