+++ Fedigardens/Modules/Profiles/Components/ProfileSheetLabel.swift
@@ -1,0 +1,35 @@
+//
+//  ProfileSheetLabel.swift
+//  Fedigardens
+//
+//  Created by Marquis Kurt on 1/28/23.
+//
+//  This file is part of Fedigardens.
+//
+//  Fedigardens is non-violent software: you can use, redistribute, and/or modify it under the terms of the CNPLv7+
+//  as found in the LICENSE file in the source code root directory or at <https://git.pixie.town/thufie/npl-builder>.
+//
+//  Fedigardens comes with ABSOLUTELY NO WARRANTY, to the extent permitted by applicable law. See the CNPL for
+//  details.
+
+import SwiftUI
+import Alice
+import EmojiText
+
+struct ProfileSheetLabel: View {
+    var profile: Account
+
+    var body: some View {
+        VStack(spacing: 4) {
+            AccountImage(author: profile)
+                .profileSize(.xxlarge)
+            EmojiText(markdown: profile.getAccountName(), emojis: profile.emojis.map { $0.remote() })
+                .font(.system(.largeTitle, design: .rounded))
+                .multilineTextAlignment(.center)
+                .bold()
+            Text("@\(profile.acct)")
+                .font(.headline)
+                .foregroundColor(.secondary)
+        }
+    }
+}
